refactor(circuit): extract branch_metadata.py to break import cycle [Clean Code: dependency direction]
`circuit_input.find_jj_order` previously did a function-local lazy
import of `_junction_order` from `circuit_utils` to mask a circular
dependency. Move both `_junction_order` and
`_capacitance_variable_for_branch` into a new dependency-free
`scqubits.core.branch_metadata` module. The five call sites
(`circuit_input`, `circuit_utils`, `symbolic_circuit`,
`symbolic_circuit_graph`, `tests/test_circuit_utils`) now import
directly from the new module.

The `circuit_utils → circuit_input` direction (used by
`assemble_circuit` for `remove_comments` etc.) survives — that's not a
cycle, just a dependency. Verified by importing both modules in either
order in a fresh interpreter.

Named `branch_metadata.py` (not `branch_types.py`) to avoid colliding
with the `BRANCH_TYPES` dict that `circuit_input.py` already exports.
